HW 3 Part B Transcription
Transcribed Part B from images to text file. Please check images for correct answers as errors in transcription are possible. 

This could be done by first checking the text file, and, if the question is incorrect on the text file, check the corresponding question in the image to see if the error is consistent.